Rename calculator package
