+++ pkg/nodes_calculator/nodes_calculator.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package calculator
+package nodes_calculator
 
 import (
 	"math"
